swap in the new js decoder, remove old
