+++ js/src/decoder/DoubleDecoder.ts
