+++ js/src/decoder/FloatDecoder.ts
