+++ js/src/decoder/IntegerDecoder.ts
