+++ js/src/decoder/PropertyDecoder.ts
@@ -1,6 +1,6 @@
 import { StreamMetadata } from '../metadata/stream/StreamMetadata';
 import { StreamMetadataDecoder } from '../metadata/stream/StreamMetadataDecoder';;
-import { Column, ScalarType } from "../../decoder/mlt_tileset_metadata_pb";
+import { Column, ScalarType } from "../metadata/mlt_tileset_metadata_pb";
 import { IntWrapper } from './IntWrapper';
 import { DecodingUtils } from './DecodingUtils';
 import { IntegerDecoder } from './IntegerDecoder';
