+++ js/src/decoder/StringDecoder.ts
