+++ js/src/index.ts
@@ -1,2 +1,2 @@
-export { MltDecoder } from './decoder2/decoder/MltDecoder';
-export { TileSetMetadata } from './decoder/mlt_tileset_metadata_pb';
+export { MltDecoder } from './decoder/MltDecoder';
+export { TileSetMetadata } from './metadata/mlt_tileset_metadata_pb';
