+++ js/src/metadata/stream/DictionaryType.ts
