+++ js/src/metadata/stream/MortonEncodedStreamMetadata.ts
