+++ js/src/metadata/stream/OffsetType.ts
