+++ js/src/metadata/stream/PhysicalStreamType.ts
