+++ js/src/metadata/stream/RleEncodedStreamMetadata.ts
