+++ js/src/metadata/stream/StreamMetadata.ts
