+++ js/src/metadata/stream/StreamMetadataDecoder.ts
