+++ js/test/unit/decoder/encodings/decodeFsst.spec.ts
@@ -1,4 +1,4 @@
-import { decodeFsst } from "../../../src/encodings/fsst";
+import { decodeFsst } from "../../../../src/encodings/fsst";
 
 describe("DecodeFsst", (): void => {
     const expectedOutput: string = "AAAAAAABBBAAACCdddddEEEEEEfffEEEEAAAAAddddCC";
